requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
aiosmtplib==2.0.2
aiosqlite==0.19.0
alembic==1.12.1
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
billiard==4.2.0
black==23.11.0
blinker==1.7.0
celery==5.3.5
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.104.1
fastapi-cache2==0.2.1
fastapi-mail==1.4.1
flake8==6.1.0
frozenlist==1.4.0
greenlet==3.0.1
gunicorn==21.2.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.2
kombu==5.3.3
Mako==1.3.0
MarkupSafe==2.1.3
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
pendulum==2.1.2
platformdirs==4.0.0
pluggy==1.3.0
prompt-toolkit==3.0.41
pycodestyle==2.11.1
pydantic==2.4.2
pydantic-settings==2.1.0
pydantic_core==2.10.1
pyflakes==3.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-dotenv==0.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pytzdata==2020.1
redis==4.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0.post1
vine==5.1.0
wcwidth==0.2.10
yarl==1.9.2
